2.3 front ref Выделен TreatmentCard в отдельный компонент.
