+++ frontend/src/components/PatientHistory/PatientCard/PatientCard.template.html
@@ -1,4 +1,4 @@
-<div class="patient-info-compact">
+<div class="patient-info-compact" v-if="store.patient">
     <!-- Заголовок с кнопками -->
     <div class="treatment-section-header">
         <h3>Персональные данные</h3>
