+++ frontend/src/components/PatientHistory/TreatmentCard/TreatmentCard.template.html
@@ -1,0 +1,124 @@
+<div class="section">
+    <div v-if="store.treatmentFormError" class="treatment-form-error">{{ store.treatmentFormError }}</div>
+    <div class="treatment-section-header">
+        <h3>Лечение</h3>
+        <div class="treatment-actions">
+            <button 
+                v-if="store.isActive && !store.editingTreatment"
+                class="btn-edit-treatment"
+                title="Редактировать лечение"
+                @click="store.startEditingTreatment()"
+            >✎</button>
+            <template v-if="store.editingTreatment">
+                <button class="btn-save-treatment" @click="store.saveTreatment()">Сохранить</button>
+                <button class="btn-cancel-treatment" @click="store.cancelEditingTreatment()">Отмена</button>
+            </template>
+        </div>
+    </div>
+    <div class="section-content">
+        <div class="info-row">
+            <div class="info-group">
+                <label>Диагноз:</label>
+                <span v-if="!store.editingTreatment">{{ store.treatment.diagnosis }}</span>
+                <input v-else type="text" v-model="store.editingTreatmentData.diagnosis" class="edit-input" />
+            </div>
+        </div>
+
+        <div class="info-row">
+            <div class="info-group">
+                <label>Код диагноза (МКБ-10):</label>
+                <span v-if="!store.editingTreatment">{{ store.treatment.diagnosisCode || '—' }}</span>
+                <input v-else type="text" v-model="store.editingTreatmentData.diagnosisCode" class="edit-input" />
+            </div>
+        </div>
+
+        <div class="info-row">
+            <div class="info-group">
+                <label>Осложнения:</label>
+                <span v-if="!store.editingTreatment">{{ store.treatment.comorbidities || '—' }}</span>
+                <input v-else type="text" v-model="store.editingTreatmentData.comorbiditiesRaw" class="edit-input" />
+            </div>
+        </div>
+
+        <div class="info-row">
+            <div class="info-group">
+                <label>Целевой диапазон МНО:</label>
+                <template v-if="!store.editingTreatment">
+                    <span>{{ store.treatment.mnoFrom }} – {{ store.treatment.mnoTo }}</span>
+                </template>
+                <template v-else>
+                    <span class="range-edit">
+                        <input type="number" step="0.1" v-model.number="store.editingTreatmentData.mnoFrom" class="edit-input small" />
+                        <span> – </span>
+                        <input type="number" step="0.1" v-model.number="store.editingTreatmentData.mnoTo" class="edit-input small" />
+                    </span>
+                </template>
+            </div>
+        </div>
+
+        <div class="info-row">
+            <div class="info-group">
+                <label>Препарат:</label>
+                <span v-if="!store.editingTreatment">{{ store.treatment.drugName || '—' }}</span>
+                <select v-else v-model.number="store.editingTreatmentData.drugId" class="edit-input">
+                    <option v-for="drug in store.allDrugs" :key="drug.id" :value="drug.id">{{ drug.nominative }}</option>
+                </select>
+            </div>
+        </div>
+
+        <div class="info-row">
+            <div class="info-group">
+                <label>Дата госпитализации:</label>
+                <span v-if="!store.editingTreatment">{{ formatDate(store.treatment.begDt) }}</span>
+                <input v-else type="date" v-model="store.editingTreatmentData.begDt" class="edit-input" />
+            </div>
+        </div>
+
+        <div class="info-row">
+            <div class="info-group">
+                <label>Плановая дата окончания:</label>
+                <span v-if="!store.editingTreatment">{{ formatDate(store.treatment.planEndDt) }}</span>
+                <input v-else type="date" v-model="store.editingTreatmentData.planEndDt" class="edit-input" />
+            </div>
+        </div>
+
+        <div class="info-row">
+            <div class="info-group">
+                <label>Комментарий:</label>
+                <span v-if="!store.editingTreatment">{{ store.treatment.comment || '—' }}</span>
+                <textarea v-else v-model="store.editingTreatmentData.treatmentComment" class="edit-textarea" rows="3"></textarea>
+            </div>
+        </div>
+
+        <template v-if="!store.isActive">
+            <div class="info-row">
+                <div class="info-group">
+                    <label>Дата завершения:</label>
+                    <span>{{ formatDate(store.treatment.realEndDt) }}</span>
+                </div>
+            </div>
+            <div class="info-row">
+                <div class="info-group">
+                    <label>Причина завершения:</label>
+                    <span>{{ store.treatment.stoppingReason || '—' }}</span>
+                </div>
+            </div>
+        </template>
+
+        <div class="info-row">
+            <div class="info-group">
+                <label>Статус:</label>
+                <span :class="store.isActive ? 'status-active' : 'status-inactive'">
+                    {{ store.isActive ? 'Активно' : 'Завершено' }}
+                </span>
+            </div>
+        </div>
+
+        <div class="info-row">
+            <div class="info-group">
+                <label>Кровотечения:</label>
+                <span>{{ store.treatment.hemorrhages ? 'есть' : 'нет' }}</span>
+            </div>
+        </div>
+    </div>
+</div>
