+++ frontend/src/components/PatientHistory/TreatmentCard/TreatmentCard.vue
@@ -1,0 +1,3 @@
+<template src="./TreatmentCard.template.html"></template>
+<script src="./TreatmentCard.script.js"></script>
+<style scoped src="./TreatmentCard.style.css"></style>
